requirements.txt
streamlit
google-genai
python-dotenv
yfinance
googlenews
requests